# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OctoManger API v2
  version: 2.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v2/config/{key}:
    get:
      operationId: getConfig
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: System configuration value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigValue"
    put:
      operationId: setConfig
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - value
              properties:
                value: {}
      responses:
        "200":
          description: Updated configuration value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemConfigValue"
...
  /api/v2/job-definitions:
    get:
      operationId: listJobDefinitions
      responses:
        "200":
          description: Job definitions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/JobDefinition"
    post:
      operationId: createJobDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobDefinitionCreateInput"
      responses:
        "201":
          description: Created job definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDefinition"
  /api/v2/job-definitions/{id}:
    get:
      operationId: getJobDefinition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job definition detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDefinition"
    patch:
      operationId: patchJobDefinition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobDefinitionPatchInput"
      responses:
        "200":
          description: Updated job definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDefinition"
    delete:
      operationId: deleteJobDefinition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted job definition
  /api/v2/job-definitions/{id}/executions:
    post:
      operationId: enqueueJobExecution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Enqueued job execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobExecution"
...
  /api/v2/job-executions/{id}:
    get:
      operationId: getJobExecution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job execution detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobExecution"
  /api/v2/job-executions/{id}/events:
    get:
      operationId: streamJobExecutionEvents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job execution SSE stream
          content:
            text/event-stream:
              schema:
                type: string
...
  /api/v2/triggers:
    get:
      operationId: listTriggers
      responses:
        "200":
          description: Triggers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Trigger"
    post:
      operationId: createTrigger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerCreateInput"
      responses:
        "201":
          description: Created trigger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerCreateResult"
  /api/v2/triggers/{id}:
    get:
      operationId: getTrigger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trigger detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trigger"
    patch:
      operationId: patchTrigger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerPatchInput"
      responses:
        "200":
          description: Updated trigger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trigger"
    delete:
      operationId: deleteTrigger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted trigger
  /api/v2/triggers/{id}/fire:
    post:
      operationId: fireTrigger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerFireInput"
      responses:
        "200":
          description: Trigger fire result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerFireResult"
  /api/v2/webhooks/{key}:
    post:
      operationId: fireWebhookTrigger
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerFireInput"
      responses:
        "200":
          description: Trigger fire result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerFireResult"
components:
  schemas:
    SystemConfigValue:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value: {}
...
    Schedule:
      type: object
      required:
        - id
        - cron_expression
        - timezone
        - next_run_at
        - enabled
      properties:
        id:
          type: integer
        cron_expression:
          type: string
        timezone:
          type: string
        next_run_at:
          type: string
          format: date-time
        enabled:
          type: boolean
    JobDefinition:
      type: object
      required:
        - id
        - key
        - name
        - plugin_key
        - action
        - input
        - enabled
      properties:
        id:
          type: integer
        key:
          type: string
        name:
          type: string
        plugin_key:
          type: string
        action:
          type: string
        input:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        schedule:
          $ref: "#/components/schemas/Schedule"
    JobExecution:
      type: object
      required:
        - id
        - job_definition_id
        - definition_key
        - definition_name
        - plugin_key
        - action
        - status
        - input
        - requested_by
        - source
      properties:
        id:
          type: integer
        job_definition_id:
          type: integer
        definition_key:
          type: string
        definition_name:
          type: string
        plugin_key:
          type: string
        action:
          type: string
        status:
          type: string
        input:
          type: object
          additionalProperties: true
        requested_by:
          type: string
        source:
          type: string
        worker_id:
          type: string
        summary:
          type: string
        error_message:
          type: string
    ScheduleInput:
      type: object
      required:
        - cron_expression
        - enabled
      properties:
        cron_expression:
          type: string
        timezone:
          type: string
        enabled:
          type: boolean
    JobDefinitionCreateInput:
      type: object
      required:
        - key
        - name
        - plugin_key
        - action
        - input
      properties:
        key:
          type: string
        name:
          type: string
        plugin_key:
          type: string
        action:
          type: string
        input:
          type: object
          additionalProperties: true
        schedule:
          $ref: "#/components/schemas/ScheduleInput"
    JobDefinitionPatchInput:
      type: object
      properties:
        name:
          type: string
        input:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        schedule:
          $ref: "#/components/schemas/ScheduleInput"
...
    Trigger:
      type: object
      required:
        - id
        - key
        - name
        - job_definition_id
        - mode
        - default_input
        - token_prefix
        - enabled
      properties:
        id:
          type: integer
        key:
          type: string
        name:
          type: string
        job_definition_id:
          type: integer
        mode:
          type: string
        default_input:
          type: object
          additionalProperties: true
        token_prefix:
          type: string
        enabled:
          type: boolean
    TriggerCreateInput:
      type: object
      required:
        - key
        - name
        - job_definition_id
        - mode
        - default_input
        - enabled
      properties:
        key:
          type: string
        name:
          type: string
        job_definition_id:
          type: integer
        mode:
          type: string
        default_input:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
    TriggerPatchInput:
      type: object
      properties:
        name:
          type: string
        mode:
          type: string
        enabled:
          type: boolean
    TriggerCreateResult:
      type: object
      required:
        - trigger
        - delivery_token
      properties:
        trigger:
          $ref: "#/components/schemas/Trigger"
        delivery_token:
          type: string
    TriggerFireInput:
      type: object
      additionalProperties: true
    TriggerFireResult:
      type: object
      required:
        - mode
        - trigger_key
      properties:
        mode:
          type: string
        trigger_key:
          type: string
        execution_id:
          type: integer
        result:
          type: object
          additionalProperties: true
        events:
          type: array
          items:
            type: object
            additionalProperties: true